修改半结构化采集公共类
